+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
  - Fields that are completely blank in the data file won't be included in the resulting map even if their definition says they can't be trimmed.
  - Added support for NAACCR reserved fields; this feature is only available for NAACCR 16 and later.
+ - This library now requires Java 8.
 
 **Changes in version 1.3.4**